requirements.txt
pytest-cov~=4.0.0
pytest==7.2.1
requests==2.28.2
PyYAML~=6.0
psycopg2-binary==2.9.5
setuptools==66.0.0
Werkzeug~=2.2.2
urllib3==1.26.14
aiohttp~=3.8.3
praw~=7.6.1
prawcore~=2.3.0
config~=0.5.1
dill~=0.3.6
configparser~=5.3.0
mock==5.0.1
PyQt6~=6.4.0
screeninfo~=0.8.1
Flask~=2.2.2
PySide6==6.4.2
pyqtdarktheme==2.1.0